A set of fixes from Plemondev (partially used)
https://github.com/opencart/opencart/pull/15333
https://github.com/opencart/opencart/pull/15397

+++ upload/admin/model/sale/order.php
@@ -887,7 +887,7 @@ class Order extends \Opencart\System\Engine\Model {
 
 		$query = $this->db->query($sql);
 
-		return (int)$query->row['total'];
+		return (float)$query->row['total'];
 	}
 
 	/**
